+++ dashboard/app.js
@@ -2,8 +2,7 @@
 
 import hljs from "https://esm.sh/highlight.js@11.10.0/lib/common";
 import htm from "https://esm.sh/htm@3.1.1";
-import { Marked, marked } from "https://esm.sh/marked@12.0.2";
-import { Component, h, render } from "https://esm.sh/preact@10.22.0";
+import { h, render } from "https://esm.sh/preact@10.22.0";
 import {
   useCallback,
   useEffect,
@@ -12,191 +11,21 @@ import {
   useState,
 } from "https://esm.sh/preact@10.22.0/hooks";
 import { MODE, TOKEN, api } from "./src/lib/api";
+import { ToastStack, appBus, showToast, toastBus } from "./src/lib/bus";
+import { ErrorBoundary, ErrorOverlay } from "./src/lib/error-boundary";
 import { fmtBytes, fmtNum, fmtPct, fmtRelativeTime, fmtUsd } from "./src/lib/format";
+import {
+  escapeHtml,
+  hlLine,
+  langFromPath,
+  renderHighlightedBlock,
+  renderMarkdownToString,
+  renderSearchReplace,
+  renderUnifiedDiff,
+} from "./src/lib/markdown";
+import { usePoll } from "./src/lib/use-poll";
 
 const html = htm.bind(h);
-
-// ---------- Markdown rendering ----------
-//
-// Rules:
-//  - GFM (tables, strikethrough, autolinks).
-//  - Code blocks with a known language → highlight.js. Auto-detect
-//    when the fence has no language. The CSS in app.css picks colors
-//    from our palette so the result reads as Reasonix's TUI, not as
-//    GitHub's.
-//  - Code blocks that look like Reasonix's edit_file SEARCH/REPLACE
-//    or a unified diff → custom diff renderer with red/green lines
-//    matching the TUI's edit-block view.
-
-function escapeHtml(s) {
-  // Defensive: marked occasionally hands the renderer `text === undefined`
-  // for empty / malformed fences (and the token-vs-positional ambiguity
-  // around v12 makes it easier to crash than not). Coerce to string
-  // instead of letting `.replace` blow up.
-  if (s == null) return "";
-  return String(s)
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-const SEARCH_REPLACE_RE = /<{7}\s*SEARCH\s*\n([\s\S]*?)\n={7}\s*\n([\s\S]*?)\n>{7}\s*REPLACE/;
-
-function renderSearchReplace(search, replace, file) {
-  const safeSearch = typeof search === "string" ? search : String(search ?? "");
-  const safeReplace = typeof replace === "string" ? replace : String(replace ?? "");
-  const oldLines = safeSearch
-    .split("\n")
-    .map((l) => `<span class="diff-line del">- ${escapeHtml(l)}</span>`)
-    .join("\n");
-  const newLines = safeReplace
-    .split("\n")
-    .map((l) => `<span class="diff-line ins">+ ${escapeHtml(l)}</span>`)
-    .join("\n");
-  const header = file ? `<span class="diff-line hunk">▸ edit ${escapeHtml(file)}</span>\n` : "";
-  return `<pre class="diff-block">${header}${oldLines}\n${newLines}</pre>`;
-}
-
-function renderUnifiedDiff(text) {
-  const safe = typeof text === "string" ? text : String(text ?? "");
-  const lines = safe
-    .split("\n")
-    .map((l) => {
-      if (l.startsWith("+++") || l.startsWith("---")) {
-        return `<span class="diff-line meta">${escapeHtml(l)}</span>`;
-      }
-      if (l.startsWith("+")) {
-        return `<span class="diff-line ins">${escapeHtml(l)}</span>`;
-      }
-      if (l.startsWith("-")) {
-        return `<span class="diff-line del">${escapeHtml(l)}</span>`;
-      }
-      if (l.startsWith("@@")) {
-        return `<span class="diff-line hunk">${escapeHtml(l)}</span>`;
-      }
-      return escapeHtml(l);
-    })
-    .join("\n");
-  return `<pre class="diff-block">${lines}</pre>`;
-}
-
-const renderer = new marked.Renderer();
-// Accept BOTH calling conventions so we don't crash if marked changes
-// shape between versions:
-//   - v12+ token object: renderer.code({ type, raw, lang, text, ... })
-//   - legacy positional:  renderer.code(text, lang, escaped)
-// Either path normalizes to `{ text, lang }`. Empty / nullable text is
-// coerced to "" (escapeHtml does the same) so an empty fence renders
-// as an empty code block rather than throwing.
-renderer.code = function reasonixCode(arg1, arg2 /* legacy lang */) {
-  let text;
-  let lang;
-  if (arg1 && typeof arg1 === "object" && !Array.isArray(arg1)) {
-    text = arg1.text;
-    lang = arg1.lang;
-  } else {
-    text = arg1;
-    lang = arg2;
-  }
-  if (text == null) text = "";
-  if (typeof text !== "string") text = String(text);
-  // Reasonix's edit_file marker block. Show as red/green diff with a
-  // small "▸ edit <file>" header lifted from the language tag (e.g.
-  // ```edit:src/foo.ts → file = src/foo.ts).
-  const sr = SEARCH_REPLACE_RE.exec(text);
-  if (sr) {
-    const file = typeof lang === "string" && lang.startsWith("edit:") ? lang.slice(5) : "";
-    return renderSearchReplace(sr[1], sr[2], file);
-  }
-  if (lang === "diff") {
-    return renderUnifiedDiff(text);
-  }
-  // Standard highlight.js path — explicit language wins, otherwise auto.
-  if (lang && typeof lang === "string" && hljs.getLanguage(lang)) {
-    try {
-      const h = hljs.highlight(text, { language: lang, ignoreIllegals: true }).value;
-      return `<pre><code class="hljs language-${lang}">${h}</code></pre>`;
-    } catch {
-      /* fall through to auto */
-    }
-  }
-  try {
-    const auto = hljs.highlightAuto(text);
-    return `<pre><code class="hljs">${auto.value}</code></pre>`;
-  } catch {
-    return `<pre><code>${escapeHtml(text)}</code></pre>`;
-  }
-};
-
-marked.use({ renderer, gfm: true, breaks: false, pedantic: false });
-
-// Separate Marked instance for the editor's markdown preview. The chat
-// renderer above does fancy SEARCH/REPLACE diff blocks and stamps every
-// code fence through hljs — useful inside an assistant message, but
-// disruptive when the user is just previewing a normal README and
-// expects standard `<pre><code>` blocks. Vanilla rendering also avoids
-// any chance our custom token-shape sniffing breaks on real markdown.
-const previewMarked = new Marked({ gfm: true, breaks: false, pedantic: false });
-previewMarked.use({
-  renderer: {
-    code(...args) {
-      const first = args[0];
-      const arg = first && typeof first === "object" ? first : { text: first, lang: args[1] };
-      const text = arg.text == null ? "" : String(arg.text);
-      const lang = typeof arg.lang === "string" ? arg.lang : "";
-      try {
-        const out =
-          lang && hljs.getLanguage(lang)
-            ? hljs.highlight(text, { language: lang, ignoreIllegals: true })
-            : hljs.highlightAuto(text);
-        const cls = lang ? `hljs language-${lang}` : "hljs";
-        return `<pre><code class="${cls}">${out.value}</code></pre>`;
-      } catch {
-        return `<pre><code>${escapeHtml(text)}</code></pre>`;
-      }
-    },
-  },
-});
-
-// usePoll: re-fetch a GET endpoint every `intervalMs`, returning
-// `{ data, error, loading, refresh }`. v0.13 swaps this for SSE.
-function usePoll(path, intervalMs = 2000) {
-  const [data, setData] = useState(null);
-  const [error, setError] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  const refresh = useCallback(async () => {
-    try {
-      const next = await api(path);
-      setData(next);
-      setError(null);
-    } catch (err) {
-      setError(err);
-    } finally {
-      setLoading(false);
-    }
-  }, [path]);
-
-  useEffect(() => {
-    let cancelled = false;
-    let timer = null;
-    const tick = async () => {
-      if (cancelled) return;
-      await refresh();
-      if (cancelled) return;
-      timer = setTimeout(tick, intervalMs);
-    };
-    tick();
-    return () => {
-      cancelled = true;
-      if (timer) clearTimeout(timer);
-    };
-  }, [refresh, intervalMs]);
-
-  return { data, error, loading, refresh };
-}
 
 // ---------- panels ----------
 
@@ -543,80 +372,7 @@ const ROLE_GLYPH = {
 
 function renderMessageBody(text) {
   if (!text) return null;
-  // marked.parse escapes raw HTML in source by default — so any `<script>`
-  // in model output gets rendered as literal text, not executed. We can
-  // safely hand the result straight to dangerouslySetInnerHTML.
-  const rendered = marked.parse(text);
-  return html`<div class="md" dangerouslySetInnerHTML=${{ __html: rendered }}></div>`;
-}
-
-// Map common file extensions to highlight.js languages.
-const LANG_BY_EXT = {
-  ts: "typescript",
-  tsx: "typescript",
-  js: "javascript",
-  jsx: "javascript",
-  mjs: "javascript",
-  cjs: "javascript",
-  py: "python",
-  rs: "rust",
-  go: "go",
-  java: "java",
-  kt: "kotlin",
-  c: "c",
-  h: "c",
-  cpp: "cpp",
-  cc: "cpp",
-  hpp: "cpp",
-  cs: "csharp",
-  swift: "swift",
-  rb: "ruby",
-  php: "php",
-  sh: "bash",
-  bash: "bash",
-  zsh: "bash",
-  fish: "bash",
-  ps1: "powershell",
-  json: "json",
-  yaml: "yaml",
-  yml: "yaml",
-  toml: "ini",
-  xml: "xml",
-  html: "xml",
-  svg: "xml",
-  css: "css",
-  scss: "scss",
-  less: "less",
-  md: "markdown",
-  sql: "sql",
-  vue: "xml",
-  svelte: "xml",
-  tex: "latex",
-  proto: "protobuf",
-  dockerfile: "dockerfile",
-};
-
-function langFromPath(path) {
-  if (!path) return null;
-  const lower = path.toLowerCase();
-  if (lower.endsWith("dockerfile")) return "dockerfile";
-  const dot = lower.lastIndexOf(".");
-  if (dot < 0) return null;
-  const ext = lower.slice(dot + 1);
-  return LANG_BY_EXT[ext] ?? null;
-}
-
-function renderHighlightedBlock(text, lang) {
-  if (!text) return "";
-  const safeLang = lang && hljs.getLanguage(lang) ? lang : null;
-  try {
-    const out = safeLang
-      ? hljs.highlight(text, { language: safeLang, ignoreIllegals: true })
-      : hljs.highlightAuto(text);
-    return `<pre class="md"><code class="hljs ${safeLang ? `language-${safeLang}` : ""}">${out.value}</code></pre>`;
-  } catch {
-    return `<pre><code>${escapeHtml(text)}</code></pre>`;
-  }
+  return html`<div class="md" dangerouslySetInnerHTML=${{ __html: renderMarkdownToString(text) }}></div>`;
 }
 
 function parseToolArgs(raw) {
@@ -980,22 +736,6 @@ function pairDiffRows(diff) {
     }
   }
   return rows;
-}
-
-// Highlight a single line via hljs in the file's language; falls back to
-// auto-detect, then escaped plain text. Always returns inline HTML safe
-// to drop into a span.
-function hlLine(text, lang) {
-  if (text == null) return "";
-  if (text === "") return "";
-  try {
-    if (lang && hljs.getLanguage(lang)) {
-      return hljs.highlight(text, { language: lang, ignoreIllegals: true }).value;
-    }
-    return hljs.highlightAuto(text).value;
-  } catch {
-    return escapeHtml(text);
-  }
 }
 
 function EditReviewModal({ modal, onResolve }) {
@@ -3774,242 +3514,6 @@ const TABS = [
     badge: null,
   },
 ];
-
-// ---------- Toast system ----------
-//
-// One Set of currently-displayed toast objects, pushed via a custom
-// DOM event so any panel can fire a toast without prop-drilling. Auto-
-// dismiss after `ttl` ms (default 3000). The stack lives at the App
-// level so toasts persist across tab switches.
-
-const toastBus = new EventTarget();
-function showToast(text, kind = "info", ttl = 3000) {
-  toastBus.dispatchEvent(new CustomEvent("toast", { detail: { text, kind, ttl } }));
-}
-
-// ---------- App-wide event bus ----------
-//
-// Two events:
-//   - "navigate-tab"  { tabId }             App switches active sidebar tab
-//   - "error"         { error, source }     global ErrorOverlay shows it full-screen
-
-const appBus = new EventTarget();
-
-// ---------- Global error capture ----------
-//
-// Three sources feed into one overlay:
-//   1. window.error              — sync exceptions, script load failures
-//   2. window.unhandledrejection — async promise rejections
-//   3. Preact ErrorBoundary       — render-time component exceptions
-//
-// All three normalize to `{ error, source, info? }` and dispatch via
-// appBus. ErrorOverlay queues the most recent and lets the user copy
-// the trace or open a pre-filled GitHub issue.
-
-function reportAppError(error, source, info) {
-  // Console-log so devtools still has the message even when the
-  // overlay is dismissed; keeps "what just broke" debuggable.
-  // eslint-disable-next-line no-console
-  console.error(`[reasonix dashboard] ${source}:`, error, info);
-  appBus.dispatchEvent(
-    new CustomEvent("error", { detail: { error, source, info, ts: Date.now() } }),
-  );
-}
-
-window.addEventListener("error", (ev) => {
-  // Resource-load errors (failing img/script) come through with no
-  // `error` object and are noisy; only surface real exceptions.
-  if (!ev.error) return;
-  reportAppError(ev.error, "window", ev.message);
-});
-
-window.addEventListener("unhandledrejection", (ev) => {
-  reportAppError(ev.reason, "promise");
-});
-
-function ToastStack() {
-  const [toasts, setToasts] = useState([]);
-  useEffect(() => {
-    const onToast = (ev) => {
-      const id = `${Date.now()}-${Math.random()}`;
-      const t = { id, ...ev.detail };
-      setToasts((prev) => [...prev, t]);
-      setTimeout(() => setToasts((prev) => prev.filter((x) => x.id !== id)), t.ttl);
-    };
-    toastBus.addEventListener("toast", onToast);
-    return () => toastBus.removeEventListener("toast", onToast);
-  }, []);
-  if (toasts.length === 0) return null;
-  return html`
-    <div class="toast-stack">
-      ${toasts.map((t) => html`<div key=${t.id} class="toast ${t.kind}">${t.text}</div>`)}
-    </div>
-  `;
-}
-
-// ---------- Error overlay ----------
-//
-// Renders a full-screen modal whenever a window error / promise
-// rejection / Preact render error fires through `appBus`. Includes a
-// "Copy details" button (clipboard) and "Report on GitHub" link with a
-// pre-filled body containing redacted environment info — the URL is
-// safe to surface (token is never embedded; just version + UA + the
-// trace itself).
-
-const REPO_URL = "https://github.com/esengine/reasonix";
-
-function buildIssueBody({ error, source, info }) {
-  const ua = typeof navigator === "object" ? navigator.userAgent : "(unknown)";
-  const errMsg = error?.message ?? String(error);
-  const stack = error?.stack ?? "(no stack)";
-  return [
-    "**What happened**",
-    "(describe what you were doing — typing, switching tabs, clicking a tool path, etc.)",
-    "",
-    "**Error**",
-    "```",
-    `${source}: ${errMsg}`,
-    info ? `info: ${info}` : null,
-    "",
-    stack,
-    "```",
-    "",
-    "**Environment**",
-    `- Reasonix: ${MODE}`,
-    `- Browser: ${ua}`,
-    `- URL: ${location.pathname} (token redacted)`,
-    "",
-    "_Reported from the local dashboard's error overlay._",
-  ]
-    .filter((l) => l !== null)
-    .join("\n");
-}
-
-function ErrorOverlay() {
-  const [err, setErr] = useState(null); // { error, source, info, ts }
-  const [copied, setCopied] = useState(false);
-
-  useEffect(() => {
-    const onError = (ev) => {
-      // Show only the latest — if a second fires while overlay is up,
-      // it replaces. Cumulative replay would be nice but for now the
-      // user can copy / file the issue with the most recent.
-      setErr(ev.detail);
-      setCopied(false);
-    };
-    appBus.addEventListener("error", onError);
-    return () => appBus.removeEventListener("error", onError);
-  }, []);
-
-  // Esc dismisses (assuming non-fatal).
-  useEffect(() => {
-    if (!err) return;
-    const onKey = (e) => {
-      if (e.key === "Escape") setErr(null);
-    };
-    window.addEventListener("keydown", onKey);
-    return () => window.removeEventListener("keydown", onKey);
-  }, [err]);
-
-  if (!err) return null;
-  const error = err.error;
-  const errMsg = error?.message ?? String(error);
-  const stack = error?.stack ?? "(no stack)";
-
-  const issueUrl = `${REPO_URL}/issues/new?title=${encodeURIComponent(`[dashboard] ${errMsg.slice(0, 80)}`)}&body=${encodeURIComponent(buildIssueBody(err))}`;
-
-  const copyDetails = async () => {
-    const body = buildIssueBody(err);
-    try {
-      await navigator.clipboard.writeText(body);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch {
-      /* clipboard blocked — user can still hit "report on GitHub" */
-    }
-  };
-
-  return html`
-    <div class="error-overlay">
-      <div class="error-overlay-card">
-        <div class="error-overlay-head">
-          <span class="error-overlay-icon">✦</span>
-          <div>
-            <div class="error-overlay-title">Something broke in the dashboard</div>
-            <div class="error-overlay-subtitle">${err.source} error · ${errMsg}</div>
-          </div>
-        </div>
-
-        <pre class="error-overlay-trace">${stack}</pre>
-
-        ${
-          err.info
-            ? html`<div class="error-overlay-info"><strong>info:</strong> ${err.info}</div>`
-            : null
-        }
-
-        <div class="error-overlay-help">
-          The TUI is unaffected — only this browser tab tripped. You can
-          dismiss and keep working, or report it so we can fix the
-          underlying cause.
-        </div>
-
-        <div class="error-overlay-actions">
-          <button class="primary" onClick=${copyDetails}>
-            ${copied ? "Copied ✓" : "Copy details"}
-          </button>
-          <a class="button" href=${issueUrl} target="_blank" rel="noopener noreferrer">
-            Report on GitHub
-          </a>
-          <button onClick=${() => setErr(null)} style="margin-left: auto;">Dismiss (Esc)</button>
-        </div>
-      </div>
-    </div>
-  `;
-}
-
-// Preact ErrorBoundary — catches render-time exceptions in the App
-// subtree and dispatches them to the error overlay instead of leaving
-// the user with a blank white page. Recovers automatically the first
-// few times so transient hiccups don't strand the user; if a panel
-// throws repeatedly we stop the loop and render a manual "Try again"
-// fallback so the page never looks blank-but-ticking.
-class ErrorBoundary extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { caught: false, lastErr: null, attempts: 0 };
-  }
-  static getDerivedStateFromError(error) {
-    return { caught: true, lastErr: error };
-  }
-  componentDidCatch(error, info) {
-    reportAppError(error, "render", info?.componentStack ?? "");
-    const attempts = (this.state.attempts ?? 0) + 1;
-    if (attempts >= 3) {
-      // Stop the auto-recover loop — the panel is genuinely broken,
-      // surface a "Try again" button instead of flickering.
-      this.setState({ attempts });
-      return;
-    }
-    setTimeout(() => this.setState({ caught: false, attempts }), 100);
-  }
-  render() {
-    if (this.state.caught) {
-      if ((this.state.attempts ?? 0) >= 3) {
-        return html`
-          <div class="boot" style="flex-direction: column; gap: 12px;">
-            <div>this panel keeps crashing — the error overlay has the trace.</div>
-            <button onClick=${() => this.setState({ caught: false, attempts: 0 })}>
-              Try again
-            </button>
-          </div>
-        `;
-      }
-      return html`<div class="boot">recovering…</div>`;
-    }
-    return this.props.children;
-  }
-}
 
 function App() {
   const [activeId, setActiveId] = useState("chat");
