+++ dashboard/src/types.d.ts
@@ -1,0 +1,18 @@
+declare module "https://esm.sh/preact@10.22.0" {
+  export * from "preact";
+}
+declare module "https://esm.sh/preact@10.22.0/hooks" {
+  export * from "preact/hooks";
+}
+declare module "https://esm.sh/htm@3.1.1" {
+  import htm from "htm";
+  export default htm;
+}
+declare module "https://esm.sh/marked@12.0.2" {
+  export * from "marked";
+}
+declare module "https://esm.sh/highlight.js@11.10.0/lib/common" {
+  import hljs from "highlight.js";
+  export default hljs;
+}
+declare module "https://esm.sh/*";
