+++ package.json
@@ -73,7 +73,6 @@
     "ink": "^5.1.0",
     "ink-spinner": "^5.0.0",
     "ink-text-input": "^6.0.0",
-    "marked": "^15.0.12",
     "picomatch": "^4.0.4",
     "react": "^18.3.1",
     "string-width": "^7.2.0",
@@ -105,6 +104,10 @@
     "@types/picomatch": "^4.0.3",
     "@types/react": "^18.3.12",
     "esbuild": "^0.21.5",
+    "highlight.js": "^11.10.0",
+    "htm": "^3.1.1",
+    "marked": "^12.0.2",
+    "preact": "^10.22.0",
     "simple-git-hooks": "^2.13.1",
     "tsup": "^8.3.5",
     "tsx": "^4.19.2",
